Remove temporary '2' versions of generic controller methods
