+++ internal/controllers/router/actuator.go
@@ -38,8 +38,8 @@ import (
 type (
 	osResourceT = routers.Router
 
-	createResourceActuator    = generic.CreateResourceActuator2[orcObjectPT, orcObjectT, filterT, osResourceT]
-	deleteResourceActuator    = generic.DeleteResourceActuator2[orcObjectPT, orcObjectT, osResourceT]
+	createResourceActuator    = generic.CreateResourceActuator[orcObjectPT, orcObjectT, filterT, osResourceT]
+	deleteResourceActuator    = generic.DeleteResourceActuator[orcObjectPT, orcObjectT, osResourceT]
 	reconcileResourceActuator = generic.ReconcileResourceActuator[orcObjectPT, osResourceT]
 	resourceReconciler        = generic.ResourceReconciler[orcObjectPT, osResourceT]
 	helperFactory             = generic.ResourceHelperFactory[orcObjectPT, orcObjectT, resourceSpecT, filterT, osResourceT]
