+++ internal/controllers/subnet/actuator.go
@@ -46,8 +46,8 @@ import (
 type (
 	osResourceT = subnets.Subnet
 
-	createResourceActuator    = generic.CreateResourceActuator2[orcObjectPT, orcObjectT, filterT, osResourceT]
-	deleteResourceActuator    = generic.DeleteResourceActuator2[orcObjectPT, orcObjectT, osResourceT]
+	createResourceActuator    = generic.CreateResourceActuator[orcObjectPT, orcObjectT, filterT, osResourceT]
+	deleteResourceActuator    = generic.DeleteResourceActuator[orcObjectPT, orcObjectT, osResourceT]
 	reconcileResourceActuator = generic.ReconcileResourceActuator[orcObjectPT, osResourceT]
 	resourceReconciler        = generic.ResourceReconciler[orcObjectPT, osResourceT]
 	helperFactory             = generic.ResourceHelperFactory[orcObjectPT, orcObjectT, resourceSpecT, filterT, osResourceT]
